backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
transformers==4.44.0
torch>=2.1.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
pydantic>=2.0.0
supabase
PyJWT
cryptography
cartesia==1.0.0
python-multipart==0.0.9
groq==0.11.0
httpx==0.27.2
